pyproject.toml
[project]
name = "wiki-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "fastembed>=0.7.4",
    "groq>=1.1.1",
    "langchain-text-splitters>=1.1.1",
    "wikipedia>=1.4.0",
]
